# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/wallet/payment-logs": {
      "get": {
        "operationId": "WalletController_getPaymentLogs",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/api/v1/wallet/recharge": {
      "post": {
        "operationId": "WalletController_rechargeWallet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RechargeWalletDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/v1/wallet/verify-payment": {
      "post": {
        "operationId": "WalletController_verifyPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/api/v1/wallet/transactions": {
      "get": {
        "operationId": "WalletController_getTransactions",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/v1/wallet/transactions/{transactionId}": {
      "get": {
        "operationId": "WalletController_getTransactionDetails",
        "parameters": [
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/v1/wallet/gifts/direct": {
      "post": {
        "operationId": "WalletController_sendDirectGift",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendDirectGiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/v1/wallet/gifts/history": {
      "get": {
        "operationId": "WalletController_getGiftHistory",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "context",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/api/v1/admin/admins/{adminId}": {
      "get": {
        "operationId": "AdminManagementController_getAdminDetails",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      },
      "patch": {
        "operationId": "AdminManagementController_updateAdmin",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      },
      "delete": {
        "operationId": "AdminManagementController_deleteAdmin",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminManagement"
        ]
      }
    },
...
    "/api/v1/admin/payments/recharge-packs/{amount}": {
      "delete": {
        "operationId": "AdminRechargePacksController_deletePack",
        "parameters": [
          {
            "name": "amount",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminRechargePacks"
        ]
      }
    },
...
    "/api/v1/registration/otp/send": {
      "post": {
        "operationId": "RegistrationController_sendOTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/v1/registration/otp/verify": {
      "post": {
        "operationId": "RegistrationController_verifyOTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/v1/registration/register": {
      "post": {
        "operationId": "RegistrationController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/v1/registration/status/ticket/{ticketNumber}": {
      "get": {
        "operationId": "RegistrationController_getStatusByTicket",
        "parameters": [
          {
            "name": "ticketNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
    "/api/v1/registration/status/phone": {
      "get": {
        "operationId": "RegistrationController_getStatusByPhone",
        "parameters": [
          {
            "name": "phoneNumber",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "countryCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RechargeWalletDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyPaymentDto": {
        "type": "object",
        "properties": {}
      },
      "SendDirectGiftDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateAdminDto": {
        "type": "object",
        "properties": {}
      },
...
      "RegisterSendOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number (without country code)",
            "example": "9876543210"
          },
          "countryCode": {
            "type": "string",
            "description": "Country code (without + sign)",
            "example": "91",
            "default": "91"
          }
        },
        "required": [
          "phoneNumber",
          "countryCode"
        ]
      },
      "RegisterVerifyOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number (without country code)",
            "example": "9876543210"
          },
          "countryCode": {
            "type": "string",
            "description": "Country code (without + sign)",
            "example": "91",
            "default": "91"
          },
          "otp": {
            "type": "string",
            "description": "6-digit OTP",
            "example": "123456"
          }
        },
        "required": [
          "phoneNumber",
          "countryCode",
          "otp"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name",
            "example": "Rajesh Kumar"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of birth (YYYY-MM-DD)",
            "example": "1990-05-15"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ],
            "example": "male"
          },
          "languagesKnown": {
            "description": "Languages known",
            "example": [
              "Hindi",
              "English",
              "Sanskrit"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "skills": {
            "description": "Skills/Expertise areas",
            "example": [
              "Vedic Astrology",
              "Numerology",
              "Tarot"
            ],
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "profilePicture": {
            "type": "string",
            "description": "Profile picture URL",
            "example": "https://example.com/profile.jpg"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number (without country code)",
            "example": "9876543210"
          },
          "countryCode": {
            "type": "string",
            "description": "Country code (without + sign)",
            "example": "91",
            "default": "91"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "rajesh@example.com"
          },
          "bio": {
            "type": "string",
            "description": "Short bio/description",
            "example": "Experienced Vedic astrologer with 10 years of practice"
          }
        },
        "required": [
          "name",
          "dateOfBirth",
          "gender",
          "languagesKnown",
          "skills",
          "profilePicture",
          "phoneNumber",
          "countryCode",
          "email"
        ]
      },
...
    }
  }
}